+++ tests/attributes/test_list.py
@@ -179,3 +179,22 @@ def test_nested_indexation(sheraf_database, persistent_type):
         assert m.mapping in conn.root()[ModelTest.table]["list"]["bar"]
         assert m.mapping in conn.root()[ModelTest.table]["list"]["baz"]
         assert [m] == list(ModelTest.search(list="foo"))
+
+
+@pytest.mark.parametrize(
+    "persistent_type", [sheraf.types.SmallList, sheraf.types.LargeList]
+)
+def test_nested_model_indexation(sheraf_database, persistent_type):
+    class Submodel(tests.UUIDAutoModel):
+        pass
+
+    class Model(tests.UUIDAutoModel):
+        submodels = sheraf.ListAttribute(
+            sheraf.ModelAttribute(Submodel),
+            persistent_type=persistent_type,
+        ).index()
+
+    with sheraf.connection(commit=True):
+        sub = Submodel.create()
+        m = Model.create(submodels=[sub])
+        assert [m] == Model.search(submodels=sub)